Rename Digraph.hpp to TmplDigraph.hpp
